exam_subject/Subject/requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-cors-headers==2.4.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
numpy==1.16.3
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
pycryptodome==3.8.0
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
raven==6.10.0
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
scipy==1.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
